Long_t -> long, Adds back some tests
- Separates out the MySpecTcl code from Core.

- Adds back some of the tests that used to live in SpecTcl directory.
  They now live in Core.

+++ main/Utility/daqdatatypes.h
@@ -64,8 +64,8 @@ typedef uint32_t UInt_t;
 typedef int16_t  Short_t;
 typedef uint16_t UShort_t;
 
-typedef int32_t  Long_t;
-typedef uint32_t ULong_t;
+typedef long  Long_t;
+typedef unsigned long ULong_t;
 
 typedef float Float_t;
 typedef double DFloat_t;
